refactor: 자동차 경주 단일 책임 원칙 리팩터링
